+++ CHANGELOG.md
@@ -26,16 +26,34 @@ Additional labels for pre-release metadata:
 * beta.x: shipped version under testing.
 * rc.x: stable release candidate.
 
+0.2.4 - unreleased
+------------------
 
-0.2.3 - Unreleased
+Changed
+* Transaction schema
+* Signed blocks
+
+Added
+* Bulk transaction
+
+
+
+0.2.3 - 25-11-2021
 ------------------
 
 Changed
 
 * Upgrade to Rust 2021
 * Bump "wasmtime" from v0.30 to v0.31.
+
+Added
+
 * Added smartcontract events to tx receipt
 * Added smartcontracts events subscription
+* Added sha256 host function
+* Added TPM2 signature
+* Added Node Monitor support
+
 
 0.2.2 - 05-11-2021
 ------------------